float128ppc: add naiive implementation of NewFromBig and add +-inf, zero
Note: the NewFromBig implementation is naiive and loses precision when
the mantissa is larger than double precision. Notably, only the high
double value is used to store the mantissa, not the low double value.

A proper implementation would store half of the mantissa in high and
half in low, adjusting the exponent such that

    high + low = f

and

    |high| >= |low|